FB-18: Implement membership renewal and history backend

+++ backend/src/main/java/lk/fat2fit/Fat2Fit/Service/ManageService.java
@@ -29,7 +29,7 @@ public class ManageService {
     private final UserRepository userRepository;
     private final ClientRepository clientRepository;
     private final ClientBodyMetricsRepository metricsRepository;
-    
+
     private User getCurrentUser() {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         return userRepository.findByEmail(email)
@@ -62,8 +62,8 @@ public class ManageService {
             builder.instructorStatus(ins.getStatus());
             if (ins.getEmployment() != null) {
                 builder.employmentType(ins.getEmployment().getEmploymentType())
-                       .workingHoursPerWeek(ins.getEmployment().getWorkingHoursPerWeek())
-                       .salary(ins.getEmployment().getSalary());
+                        .workingHoursPerWeek(ins.getEmployment().getWorkingHoursPerWeek())
+                        .salary(ins.getEmployment().getSalary());
             }
         }
 
@@ -156,7 +156,7 @@ public class ManageService {
 
     // ── Admin / Instructor: edit a client's details ────────────────────────────
 
-    public ResponseEntity<?> editClient(int clientId, UserEditRequest req) {
+    public ResponseEntity<?> editClient(Long clientId, UserEditRequest req) {
         Optional<Client> clientOpt = clientRepository.findById(clientId);
         if (clientOpt.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Client not found.");
@@ -180,7 +180,7 @@ public class ManageService {
 
     // ── Admin / Instructor: get client body metrics ───────────────────────────
 
-    public ResponseEntity<?> getClientMetrics(int clientId) {
+    public ResponseEntity<?> getClientMetrics(Long clientId) {
         if (!clientRepository.existsById(clientId)) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Client not found.");
         }
@@ -191,7 +191,7 @@ public class ManageService {
 
     // ── Admin / Instructor: save client body metrics ──────────────────────────
 
-    public ResponseEntity<?> saveClientMetrics(int clientId, ClientMetricsRequest req) {
+    public ResponseEntity<?> saveClientMetrics(Long clientId, ClientMetricsRequest req) {
         Optional<Client> clientOpt = clientRepository.findById(clientId);
         if (clientOpt.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Client not found.");
@@ -203,15 +203,24 @@ public class ManageService {
                         .fitnessGoals(new HashSet<>())
                         .build());
 
-        if (req.getWeightKg() != null)       metrics.setWeightKg(req.getWeightKg());
-        if (req.getHeightCm() != null)       metrics.setHeightCm(req.getHeightCm());
-        if (req.getHipSizeCm() != null)      metrics.setHipSizeCm(req.getHipSizeCm());
-        if (req.getBreastSizeCm() != null)   metrics.setBreastSizeCm(req.getBreastSizeCm());
-        if (req.getWaistSizeCm() != null)    metrics.setWaistSizeCm(req.getWaistSizeCm());
-        if (req.getArmSizeCm() != null)      metrics.setArmSizeCm(req.getArmSizeCm());
-        if (req.getShoulderSizeCm() != null) metrics.setShoulderSizeCm(req.getShoulderSizeCm());
-        if (req.getButtSizeCm() != null)     metrics.setButtSizeCm(req.getButtSizeCm());
-        if (req.getFitnessGoals() != null)   metrics.setFitnessGoals(req.getFitnessGoals());
+        if (req.getWeightKg() != null)
+            metrics.setWeightKg(req.getWeightKg());
+        if (req.getHeightCm() != null)
+            metrics.setHeightCm(req.getHeightCm());
+        if (req.getHipSizeCm() != null)
+            metrics.setHipSizeCm(req.getHipSizeCm());
+        if (req.getBreastSizeCm() != null)
+            metrics.setBreastSizeCm(req.getBreastSizeCm());
+        if (req.getWaistSizeCm() != null)
+            metrics.setWaistSizeCm(req.getWaistSizeCm());
+        if (req.getArmSizeCm() != null)
+            metrics.setArmSizeCm(req.getArmSizeCm());
+        if (req.getShoulderSizeCm() != null)
+            metrics.setShoulderSizeCm(req.getShoulderSizeCm());
+        if (req.getButtSizeCm() != null)
+            metrics.setButtSizeCm(req.getButtSizeCm());
+        if (req.getFitnessGoals() != null)
+            metrics.setFitnessGoals(req.getFitnessGoals());
         metrics.setOtherGoalSpecification(req.getOtherGoalSpecification());
 
         metricsRepository.save(metrics);
